# plumbing: format/packfile, performance optimizations for reading large commit histories