We have recently evaluated using [Gitea](https://gitea.io/en-us/) as our Git hosting solution. It turned out to have quite a few performance bottlenecks stemming from the fact that it invokes the `git` command line tool for all operations. In one of the tracking issues go-git was suggested as one way of speeding it up. Over the weekend I prototyped computing the directory listing with last commit for each file on top of go-git (https://gist.github.com/filipnavara/8e6fdf980130d6ca120bfda4c25481e9). The speed improvement seemed promising enough that yesterday I proceeded further and reimplemented part of Gitea to use go-git (https://github.com/filipnavara/gitea/tree/perf-read). Overall I am seeing about 10x improvement on page load times for directory listings on our repository.

As part of the experiment I hit couple of bottlenecks in the go-git itself and I think it would make sense to resolve them. The proposed solutions may not necessarily be the best for general use, but I would like to spark a discussion about it at least.

1. Profiling the I/O showed a lot of `Seek(0, io.SeekCurrent)` calls that were unnecessary since the offset was either not used and thrown away or already known in the code path. This accounted for up to 5% of the history traversal.
_Proposed solution:_ I added `SeekObjectHeader` to `packfile/scanner` that combines seeking in the file and then reading the object header at that position into one call. This should not adversly affect any workloads. Related to that I found that `Seek(<non-zero integer>, io.SeekCurrent)` was implemented incorrectly. It is never used in the code base, but should probably be fixed nevertheless.

2. Iterating over commits using the `repository.log`, `object.NewCommitIterCTime` or other iterators always loads the commit objects from disk for loose objects. Since we examined the commit parents inside the iterator we were essentially loading all the commits from disk twice. It is possible to rewrite our algorithms/loops to avoid that, but it seemed that the use case could be better serviced within the library.
_Proposed solution:_ Add small cache for the loaded loose objects. It is more efficient to keep couple of them in the memory instead of reading them from disk. This could probably be tweaked to cache only small objects where the I/O and syscalls are going to be most expensive compared to memory cost.

3. Iterating deep in the history eventually started hitting the packfiles. The profiler and Process Monitor showed that there was significant time spent in `OpenFile` calls. Most of these `OpenFile` calls failed because the objects were stored in the packfiles and not as loose objects.
_Proposed solution:_ If the packfile indexes are already loaded then prefer them to loose object lookup. For our use case it improves the hit rate to almost 100% and avoids the unnecessary I/O calls. The indexes are already loaded in memory and should be cheap to check compared to hitting syscalls and disk. It may not be too efficient if many packfiles are present, but then the probability of objects being in the packfile is higher anyway. This simple optimization improved our algorithm performance by nearly 30%.

4. The I/O patterns for reading small objects from packfiles was quite erratic. It resulted in almost 3x as much disk reads performed than necessary (checked by `timeit` tool) and it often loaded 4k buffered blocks only to seek back in the file and read an overlapping block again (seen in Process Monitor on Windows). It turned out to be caused by the way `FSObject` is implemented. There's potential for significant improvement.
_Proposed solution:_ For small objects use `MemoryObject` instead and read the content right away since it's already buffered in memory. I chose an arbitrary threshold of 16 Kb, which may not be ideal. Something like (4Kb - header size) would be a conservative value that would work for our use case too. 

5. Pack file index reading is done in 4Kb chunks, but it is mostly sequential reads that could easily use bigger buffer. This is not addressed within the PR.