@filipnavara 2018-09-24
Apparently the "Improve performace for reading many small objects from packfile by letting the I/O buffers be used better." change (#4) is triggering some unwanted behavior with cache. I am looking into it.