@filipnavara 2018-09-26
@mcuadros I don't have any scientific benchmarks (being a .NET guy I miss a tool like BenchmarkDotNet). For my specific use case (roughly the GIST link in the original description) on our repository with ~12 year of history, 3Gb of data and ~30000 commits it sped up my directory listing algorithm from 4s to 1.6s. It roughly corresponds to the reduced I/O calls shown by `timeit` and `go tool ppref`. Further improvements are possible, but this moved `go-git` for this specific use case into a usable territory in comparison to `git` command line. Now the profiler shows that the I/O is roughly 29% of the run time compared to 78% before the changes.

I am toying with some ideas how to make the commit graph traversal even faster and avoid a lot of parsing of information that is thrown away. Notably I have written a parser for the Git 2.18+ commit-graph files. It is very much work in progress and nothing quite ready to share yet.