@jfontan 2018-10-01
I've tested [borges](https://github.com/src-d/borges) that is a tool that downloads and packs repositories in siva files. To do this it has to go through all the commits to repack the repos. Most of the times are lower and system time (`Stime`) is consistently smaller. Only the last repo takes more time than before. Not sure why.

```
#### Comparing /root/borges/borges-4.7.0 - /root/borges/borges-4.7.0-pr-963 ####
## Repo cangallo ##
Memory: 25528 -> 26780 (4.904418677530555), true
Wtime: 849.703176ms -> 889.788468ms (4.717564101466887), true
Stime: 140ms -> 170ms (21.428571428571427), false
Utime: 1.05s -> 920ms (-12.380952380952381), true
FileSize: 116541 -> 120423 (3.3310165521147064), true
## Repo octoprint-tft ##
Memory: 60824 -> 65160 (7.1287649611995265), true
Wtime: 8.727670689s -> 8.487030095s (-2.757214411209323), true
Stime: 710ms -> 550ms (-22.535211267605636), true
Utime: 9.61s -> 9.55s (-0.6243496357960457), true
FileSize: 2950179 -> 2956207 (0.20432658492925346), true
## Repo upsilon ##
Memory: 630096 -> 628164 (-0.30661994362763767), true
Wtime: 2m18.403927785s -> 2m11.427576858s (-5.040572936511767), true
Stime: 12.39s -> 10.1s (-18.48264729620662), true
Utime: 2m32.23s -> 2m25.9s (-4.158181698745319), true
FileSize: 100912676 -> 101071233 (0.1571229763047806), true
## Repo numpy ##
Memory: 1168112 -> 1232408 (5.504266714150698), true
Wtime: 7m5.497781807s -> 7m37.625167174s (7.550541210946323), true
Stime: 54.85s -> 46.12s (-15.916134913400182), true
Utime: 9m13.03s -> 10m35.02s (14.825597164710775), false
FileSize: 96717584 -> 110103219 (13.83991870599249), false
```