@filipnavara 2018-10-01
The slow down could be caused by something I didn't think of. There are two things that could result in worse performance in some scenarios:

1) When there is a lot of pack files and someone tries to look up loose object by hash. In that case the in-memory query with object hash across all loaded pack file indexes could be more expensive than making a syscall and hitting an OS disk cache.
2) If the loose objects are inadvertently spilling too much of the object cache and slow-down later delta lookups.

I assumed that either of those scenarios are unlikely to be hit. For 1) my thinking is that the more packfiles there are the larger is the chance that the object is stored in them. And in the unlikely case that it's not then the hash table lookup should still be cheap. For 2) I assumed that at certain point in time one will either be hitting mostly loose objects or mostly objects in pack files. Not many algorithms or traversals are randomly accessing the objects in such a fashion that accesses to loose objects and pack file objects would interleave too much.

If the Borges scenario could be reproduced on my machine I can try to pinpoint the specific optimization (commit in this PR) that causes the slow down and work on it.