@filipnavara 2018-10-01
I am not quite sure that I followed the right steps to reproduce your scenario. I created `repos.txt` file with one line: `https://github.com/numpy/numpy.git`. Then I built borges cli straight from GIT and executed `./borges pack --log-level=debug repos.txt`. I did that three times without the patch and three times with the patch (against `master`, not `4.7`). I took the time between "remote repository cloned" and "repository processed". The times I see are 2m56s with my patch, 2m59s with master. The difference is kinda in the margin of an error. However, the number are so much smaller than yours that I feel that I am actually testing a wrong thing. This is on a MacBook Pro 2017 model under macOS 10.14 with pretty speedy SSD.