@jfontan 2018-10-02
I'm testing in a slow machine (Intel(R) Atom(TM) CPU  C2550  @ 2.40GHz) with a network attached disk. That's why the numbers are that different. The difference in my machine with that repo is not that big. Around 7%. I was just curious why it could be as looking at the code it seems it should be faster.

The repos are testing are local to the machine where I'm doing the test and only have one packfile.