@ajnavarro 2018-11-16
I did some really simple tests using this PR to compare against the master branch. I used this script on [kubernetes repository](https://github.com/kubernetes/kubernetes): 

```go
package main

import (
    "fmt"
    "io"

    "gopkg.in/src-d/go-git.v4"
    . "gopkg.in/src-d/go-git.v4/_examples"
)

func main() {
    r, err := git.PlainOpen(".git")
    CheckIfError(err)

    iter, err := r.CommitObjects()
    CheckIfError(err)
    for {
        c, err := iter.Next()
        if err == io.EOF {
            break
        }

        CheckIfError(err)
        fmt.Println(c.Hash.String())
    }

    iter.Close()
}
```

Instead of taking ~20secs it is taking ~1m30s.

Maybe this can be related with no usage of the delta cache on someplace.