@filipnavara 2018-11-19
Just got back home after couple of weeks on vacation. As soon as I dig through my piled up work I will try to reproduce your benchmark and investigate it.