@smola 2018-11-19
When not interested in dangling commits, this will speed things up: https://github.com/src-d/go-git/issues/1024