@filipnavara 2018-11-26
@ajnavarro I run a profiler on it and tracked it to one of my changes. The logic for pre-loading of small objects was incorrectly applied also on deltas, which resulted in more file seeking and reading than intended, ie. exact opposite of what the optimisation tried to achieve.

I can't quite reproduce the numbers you have, but the fix reduced it from ~1m40s to ~30s on my machine (with profiling enabled).