@ajnavarro 2018-11-27
Hi @filipnavara ,

Now, executing the script I provided above, I'm having the following results:

```
09:36 $ time /tmp/count-commits-2 | wc -l
80313

real	0m20.012s
user	0m19.432s
sys	0m1.480s
```

Thanks a lot to take care of this!