@smola 2018-11-27
@jfontan @ajnavarro What about borges? Does it run faster now with latest changes?