@jfontan 2018-11-27
@smola Now borges is a little bit faster:

```
#### Comparing /root/borges/borges-4.8.0 - /root/borges/borges-4.8.0-pr-963 ####
## Repo cangallo ##
Memory: 24044 -> 24848 (3.343869572450507), true
Wtime: 783.30907ms -> 715.507071ms (-8.655842450541266), true
Stime: 130ms -> 150ms (15.384615384615385), false
Utime: 1.06s -> 990ms (-6.60377358490566), true
FileSize: 116333 -> 120157 (3.2871154358608474), true
## Repo octoprint-tft ##
Memory: 58632 -> 52472 (-10.506208213944603), true
Wtime: 8.633554855s -> 8.539363412s (-1.0909925816415051), true
Stime: 820ms -> 570ms (-30.48780487804878), true
Utime: 9.55s -> 9.28s (-2.8272251308900525), true
FileSize: 2950139 -> 2955791 (0.19158419315157693), true
## Repo upsilon ##
Memory: 628556 -> 613224 (-2.439241690477857), true
Wtime: 2m17.845137743s -> 2m11.257079352s (-4.779318660686348), true
Stime: 11.86s -> 11.02s (-7.082630691399664), true
Utime: 2m32.1s -> 2m25.09s (-4.608809993425378), true
FileSize: 100906120 -> 101068735 (0.16115474462797696), true
## Repo numpy ##
Memory: 1131220 -> 1180632 (4.3680274394017085), true
Wtime: 7m5.332720436s -> 6m56.547470917s (-2.0655005121624352), true
Stime: 55.89s -> 51.97s (-7.01377706208624), true
Utime: 9m11.83s -> 9m39.34s (4.985230958809779), true
FileSize: 96709754 -> 110124947 (13.871602858177056), false
```